backend/ai-docs-chat/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.12.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
dnspython==2.8.0
python-docx==1.1.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.76
langchain-groq==0.2.0
langchain-huggingface==0.3.1
langchain-pinecone==0.2.12
langchain-text-splitters==0.3.11
langsmith==0.4.31
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.1
openai
langchain-openai
orjson==3.11.3
packaging==25.0
passlib==1.7.4
pinecone
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.37.0
# uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
zstandard==0.25.0
huggingface-hub==0.35.1
bcrypt>=3.2.0

